services/api/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.18
asyncpg==0.27.0
alembic==1.11.1
python-jose==3.3.0
passlib[argon2]==1.7.4
pydantic==2.5.1
email-validator==2.0.0
aioredis==2.0.1
psycopg2-binary==2.9.12
prometheus_client==0.16.0
httpx==0.24.1
pytest-asyncio==0.22.2
pytest==7.4.2
psycopg2==2.9.12
pytest-mock==3.11.1
pdfplumber==0.9.0
python-docx==0.8.11
spacy==3.5.0
minio==7.1.15
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.10.1
argon2-cffi>=21.3.0
huggingface-hub>=0.16.0
jinja2==3.1.2
aiosmtplib>=5.0.0
redis==5.0.1
